Update IHero interface for power type change
Changed the IHero interface to use uint16 for the power property in getHeroProperties and mintFromAltar functions, reflecting contract optimizations for gas savings. Added deployment analysis documentation for version v1.3.7.0.

+++ contracts/current/interfaces/interfaces.sol
@@ -157,8 +157,8 @@ interface IHero {
     
     // Cross-contract functions
     function ownerOf(uint256 tokenId) external view returns (address);
-    function getHeroProperties(uint256 tokenId) external view returns (uint8 rarity, uint256 power);
-    function mintFromAltar(address to, uint8 rarity, uint256 power) external returns (uint256);
+    function getHeroProperties(uint256 tokenId) external view returns (uint8 rarity, uint16 power);
+    function mintFromAltar(address to, uint8 rarity, uint16 power) external returns (uint256);
     function burnFromAltar(uint256 tokenId) external;
     function safeTransferFrom(address from, address to, uint256 tokenId) external;
     function setApprovalForAll(address operator, bool approved) external;
